+++ ontask/dataops/forms/dataframeupload.py
@@ -9,7 +9,7 @@ from django.conf import settings
 from smart_open import smart_open
 
 from ontask.dataops.pandas import create_db_engine
-from ontask.models import SQLConnection
+from ontask.models import SQLConnection, AthenaConnection
 
 
 def _process_object_column(data_frame: pd.DataFrame) -> pd.DataFrame:
@@ -257,6 +257,39 @@ def load_df_from_sqlconnection(
 
     # Try to fetch the data
     data_frame = pd.read_sql_table(conn_item.db_table, db_engine)
+
+    # Strip white space from all string columns and try to convert to
+    # datetime just in case
+    return _process_object_column(data_frame)
+
+
+def load_df_from_athenaconnection(
+    conn_item: AthenaConnection,
+    table_name: Optional[str] = None,
+) -> pd.DataFrame:
+    """Load a DF from an Athena connection.
+
+    from pyathena import connect
+    from pyathena.pandas_cursor import PandasCursor
+
+    cursor = connect(aws_access_key_id='YOUR_ACCESS_KEY_ID',
+               aws_secret_access_key='YOUR_SECRET_ACCESS_KEY',
+               s3_staging_dir='s3://YOUR_S3_BUCKET/path/to/',
+               region_name='us-west-2',
+               cursor_class=PandasCursor).cursor()
+
+    df = cursor.execute("SELECT * FROM many_rows").as_pandas()
+    print(df.describe())
+    print(df.head())
+
+    :param conn_item: AthenaConnection object with the connection parameters.
+
+    :param table_name: table name
+
+    :return: Data frame or raise an exception.
+    """
+    # TODO: Implement this function
+    data_frame = pd.DataFrame()
 
     # Strip white space from all string columns and try to convert to
     # datetime just in case