+++ ontask/dataops/views/athenaupload.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Function to upload a data frame from an existing SQL connection object."""
+"""Function to upload a data frame from an Athena connection object."""
 
 from typing import Optional
 
@@ -14,19 +14,20 @@ from django.utils.translation import ugettext_lazy as _
 from ontask import OnTaskDataFrameNoKey
 from ontask.core.decorators import get_workflow
 from ontask.core.permissions import is_instructor
-from ontask.dataops.forms import SQLRequestPassword, load_df_from_sqlconnection
+from ontask.dataops.forms import (
+    AthenaRequestTable, load_df_from_athenaconnection)
 from ontask.dataops.pandas import store_temporary_dataframe, verify_data_frame
-from ontask.models import SQLConnection, Workflow
+from ontask.models import AthenaConnection, Workflow
 
 
 @user_passes_test(is_instructor)
 @get_workflow()
-def sqlupload_start(
+def athenaupload_start(
     request: HttpRequest,
     pk: int,
     workflow: Optional[Workflow] = None,
 ) -> HttpResponse:
-    """Load a data frame using a SQL connection.
+    """Load a data frame using an Athena connection.
 
     The four step process will populate the following dictionary with name
     upload_data (divided by steps in which they are set
@@ -42,54 +43,47 @@ def sqlupload_start(
     step_1: URL name of the first step
 
     :param request: Web request
-    :param pk: primary key of the SQL conn used
+    :param pk: primary key of the Athena conn used
     :return: Creates the upload_data dictionary in the session
     """
-    conn = SQLConnection.objects.filter(pk=pk).first()
+    conn = AthenaConnection.objects.filter(pk=pk).first()
     if not conn:
-        return redirect('dataops:sqlconns_instructor_index_instructor_index')
+        return redirect(
+            'dataops:athenaconns_instructor_index_instructor_index')
 
     form = None
-    password = conn.db_password
-    if not password:
-        # The connection needs a password  to operate
-        form = SQLRequestPassword(request.POST or None)
+    table_name = conn.table_name
+    if not table_name:
+        # The connection needs a table (not given upon definition)
+        form = AthenaRequestTable(request.POST or None)
 
     context = {
         'form': form,
         'wid': workflow.id,
-        'dtype': 'SQL',
-        'dtype_select': _('SQL connection'),
+        'dtype': 'Athena',
+        'dtype_select': _('Athena connection'),
+        'connection': conn,
+        'table_name': table_name,
         'valuerange': range(5) if workflow.has_table() else range(3),
-        'prev_step': reverse('dataops:sqlconns_instructor_index'),
-        'conn_type': conn.conn_type,
-        'conn_driver': conn.conn_driver,
-        'db_user': conn.db_user,
-        'db_passwd': _('<PROTECTED>') if password else '',
-        'db_host': conn.db_host,
-        'db_port': conn.db_port,
-        'db_name': conn.db_name,
-        'db_table': conn.db_table}
+        'prev_step': reverse('dataops:athenaconns_instructor_index')}
 
     if request.method == 'POST' and (not form or form.is_valid()):
-        if form:
-            password = form.cleaned_data['password']
+        if not table_name:
+            table_name = form.cleaned_data['table_name']
 
-        # Process SQL connection using pandas
+        # Process Athena connection using pandas
         try:
-            data_frame = load_df_from_sqlconnection(
-                conn,
-                password)
+            data_frame = load_df_from_athenaconnection(conn, table_name)
             # Verify the data frame
             verify_data_frame(data_frame)
         except OnTaskDataFrameNoKey as exc:
             messages.error(request, str(exc))
-            return render(request, 'dataops/sqlupload_start.html', context)
+            return render(request, 'dataops/athenaupload_start.html', context)
         except Exception as exc:
             messages.error(
                 request,
                 _('Unable to obtain data: {0}').format(str(exc)))
-            return render(request, 'dataops/sqlupload_start.html', context)
+            return render(request, 'dataops/athenaupload_start.html', context)
 
         # Store the data frame in the DB.
         try:
@@ -102,7 +96,7 @@ def sqlupload_start(
                 None,
                 _('The data from this connection cannot be processed.'),
             )
-            return render(request, 'dataops/sqlupload_start.html', context)
+            return render(request, 'dataops/athenaupload_start.html', context)
 
         # Dictionary to populate gradually throughout the sequence of steps. It
         # is stored in the session.
@@ -111,10 +105,10 @@ def sqlupload_start(
             'column_types': frame_info[1],
             'src_is_key_column': frame_info[2],
             'step_1': reverse(
-                'dataops:sqlupload_start',
+                'dataops:athenaupload_start',
                 kwargs={'pk': conn.id}),
         }
 
         return redirect('dataops:upload_s2')
 
-    return render(request, 'dataops/sqlupload_start.html', context)
+    return render(request, 'dataops/athenaupload_start.html', context)