+++ ontask/dataops/views/upload.py
@@ -19,7 +19,7 @@ from ontask.dataops.forms import SelectColumnUploadForm, SelectKeysForm
 from ontask.dataops.pandas import load_table, perform_dataframe_upload_merge
 from ontask.dataops.pandas.dataframe import store_workflow_table
 from ontask.dataops.sql import table_queries
-from ontask.models import Log, SQLConnection, Workflow
+from ontask.models import Log, SQLConnection, Workflow, AthenaConnection
 from ontask.workflow.access import store_workflow_in_session
 
 
@@ -43,8 +43,8 @@ def uploadmerge(
         'dataops/uploadmerge.html',
         {
             'valuerange': range(5) if workflow.has_table() else range(3),
-            'sql_enabled': SQLConnection.objects.count() > 0
-        })
+            'sql_enabled': SQLConnection.objects.count() > 0,
+            'athena_enabled': AthenaConnection.objects.count() > 0})
 
 
 @user_passes_test(is_instructor)