+++ ontask/static/dataops/js/dataops.js
@@ -49,13 +49,13 @@ $(function () {
   $("#sql-connections").on("click", ".js-sqlconn-add", loadForm);
   $("#modal-item").on("submit", ".js-sqlconn-add-form", saveForm);
 
-  $("#sqlconn-admin-table").on("click", ".js-sqlconn-edit", loadForm);
+  $("#conn-admin-table").on("click", ".js-sqlconn-edit", loadForm);
   $("#modal-item").on("submit", ".js-sqlconn-edit-form", saveForm);
 
-  $("#sqlconn-admin-table, #modal-item").on("click", ".js-sqlconn-delete", loadForm);
+  $("#conn-admin-table, #modal-item").on("click", ".js-sqlconn-delete", loadForm);
   $("#modal-item").on("submit", ".js-sqlconn-delete-form", saveForm);
 
-  $("#sqlconn-admin-table, #modal-item").on("click", ".js-sqlconn-clone", loadForm);
+  $("#conn-admin-table, #modal-item").on("click", ".js-sqlconn-clone", loadForm);
   $("#modal-item").on("submit", ".js-sqlconn-clone-form", saveForm);
 
   // Athena connection add, edit, delete and clone
@@ -64,13 +64,13 @@ $(function () {
   $("#athena-connections").on("click", ".js-athenaconn-add", loadForm);
   $("#modal-item").on("submit", ".js-athenaconn-add-form", saveForm);
 
-  $("#athenaconn-admin-table").on("click", ".js-athenaconn-edit", loadForm);
+  $("#conn-admin-table").on("click", ".js-athenaconn-edit", loadForm);
   $("#modal-item").on("submit", ".js-athenaconn-edit-form", saveForm);
 
-  $("#athenaconn-admin-table, #modal-item").on("click", ".js-athenaconn-delete", loadForm);
+  $("#conn-admin-table, #modal-item").on("click", ".js-athenaconn-delete", loadForm);
   $("#modal-item").on("submit", ".js-athenaconn-delete-form", saveForm);
 
-  $("#athenaconn-admin-table, #modal-item").on("click", ".js-athenaconn-clone", loadForm);
+  $("#conn-admin-table, #modal-item").on("click", ".js-athenaconn-clone", loadForm);
   $("#modal-item").on("submit", ".js-athenaconn-clone-form", saveForm);
 
   // Toggle plugin is_enabled