+++ test/__init__.py
@@ -694,7 +694,7 @@ class OnTaskLiveTestCase(LiveServerTestCase):
             '//*[@id="ontask-base-admin"]',
             'SQL Connections')
         WebDriverWait(self.selenium, 10).until(
-            EC.presence_of_element_located((By.ID, 'sqlconn-admin-table'))
+            EC.presence_of_element_located((By.ID, 'conn-admin-table'))
         )
         WebDriverWait(self.selenium, 10).until_not(
             EC.visibility_of_element_located((By.ID, 'div-spinner'))
@@ -706,7 +706,7 @@ class OnTaskLiveTestCase(LiveServerTestCase):
             '//*[@id="ontask-base-admin"]',
             'Athena Connections')
         WebDriverWait(self.selenium, 10).until(
-            EC.presence_of_element_located((By.ID, 'athenaconn-admin-table'))
+            EC.presence_of_element_located((By.ID, 'conn-admin-table'))
         )
         WebDriverWait(self.selenium, 10).until_not(
             EC.visibility_of_element_located((By.ID, 'div-spinner'))
@@ -790,6 +790,19 @@ class OnTaskLiveTestCase(LiveServerTestCase):
         WebDriverWait(self.selenium, 10).until(
             EC.presence_of_element_located(
                 (By.XPATH, '//div[@id = "sql-connections"]')
+            )
+        )
+
+    def go_to_athena_upload_merge(self):
+        self.go_to_upload_merge()
+
+        # Goto SQL option
+        self.selenium.find_element_by_xpath(
+            '//table[@id="dataops-table"]//a[normalize-space()="Athena '
+            'Connection"]').click()
+        WebDriverWait(self.selenium, 10).until(
+            EC.presence_of_element_located(
+                (By.XPATH, '//div[@id = "athena-connections"]')
             )
         )