+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+- New functionality to upload/merge data using an Amazon Athena connection (Issue #166)
+
 - New functionality to execute actions incrementally over a time window (Issue #160)
 
 - New action type: personalized rubric. Write a rubric, extend each cell with a feedback paragraph, collect data about the level of attainment and create a personalized email with the appropriate messages for each student.
@@ -9,6 +11,10 @@
 - New button to export a subset of actions from a workflow in a single step.
 
 - Link to the table views in the top menu
+
+### Changed
+
+- Revamped the logs created by the platform so that they have a uniform structure
 
 ### Fixed